requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.9.0
black==25.1.0
blis==1.3.0
camelot-py==1.0.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
cymem==2.0.11
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@b15f64ec4429e23a148972175a0207c5a9ab84cf
effdet==0.4.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
et-xmlfile==2.0.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
fvcore==0.1.5.post20221221
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
iopath==0.1.9
jinja2==3.1.6
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language-data==1.3.0
layoutparser @ git+https://github.com/Layout-Parser/layout-parser.git@04e28168d820eea3a1ff1e098078323e7b48648b
marisa-trie==1.2.1
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
mypy-extensions==1.1.0
networkx==3.5
numpy==2.3.1
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
openai==1.57.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pdf2image==1.17.0
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pip==25.1.1
platformdirs==4.3.8
portalocker==3.2.0
preshed==3.0.10
protobuf==6.31.1
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pymupdf==1.24.14
pyparsing==3.2.3
pypdf==3.17.4
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-doctr==0.12.0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
rapidfuzz==3.13.0
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
thinc==8.3.6
timm==1.0.16
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
upstash-redis==1.4.0
urllib3==2.5.0
validators==0.35.0
wasabi==1.1.3
weasel==0.4.1
werkzeug==3.1.3
wrapt==1.17.2
yacs==0.1.8
